Add ServerName to Disk's ColumnDefs
